@@ -17,7 +17,7 @@
     "yarn": "1.x"
   },
   "devDependencies": {
-    "@angular-devkit/build-optimizer": "^0.11.0",
+    "@angular-devkit/build-optimizer": "^0.11.1",
     "@angular/animations": "^7.1.1",
     "@angular/common": "^7.1.1",
     "@angular/compiler": "^7.1.1",
@@ -29,9 +29,9 @@
     "@angular/platform-browser-dynamic": "^7.1.1",
     "@angular/router": "^7.1.1",
     "@angular/upgrade": "^7.1.1",
-    "@babel/core": "^7.1.6",
-    "@babel/plugin-syntax-dynamic-import": "^7.0.0",
-    "@babel/preset-env": "^7.1.6",
+    "@babel/core": "^7.2.0",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
+    "@babel/preset-env": "^7.2.0",
     "@babel/register": "^7.0.0",
     "@fortawesome/angular-fontawesome": "^0.3.0",
     "@fortawesome/fontawesome": "^1.1.8",
@@ -40,26 +40,27 @@
     "@fortawesome/free-regular-svg-icons": "^5.5.0",
     "@fortawesome/free-solid-svg-icons": "^5.5.0",
     "@ng-bootstrap/ng-bootstrap": "^4.0.0",
-    "@ngtools/webpack": "^7.1.0",
+    "@ngtools/webpack": "^7.1.1",
     "@types/express": "^4.16.0",
     "@types/ioredis": "^4.0.4",
-    "@types/jquery": "^3.3.22",
-    "@types/lodash": "^4.14.118",
-    "@types/marked": "^0.4.2",
-    "@types/mongodb": "^3.1.15",
-    "@types/node": "^10.12.11",
+    "@types/jquery": "^3.3.23",
+    "@types/lodash": "^4.14.119",
+    "@types/marked": "^0.5.0",
+    "@types/mongodb": "^3.1.17",
+    "@types/node": "^10.12.12",
     "@types/sanitize-html": "^1.18.2",
     "@types/serve-index": "^1.7.29",
-    "@types/socket.io": "^2.1.0",
+    "@types/socket.io": "^2.1.1",
     "@types/socket.io-client": "^1.4.32",
     "@types/socket.io-redis": "^1.0.25",
-    "@types/three": "^0.93.10",
+    "@types/three": "^0.93.12",
     "@types/uuid": "^3.4.4",
     "@types/webpack": "^4.4.20",
     "@types/webpack-env": "^1.13.6",
     "awesome-typescript-loader": "^5.2.1",
     "babel-loader": "^8.0.4",
     "bootstrap": "^4.1.3",
+    "breakpoints-js": "^1.0.6",
     "browser-sync": "^2.26.3",
     "css-loader": "^1.0.1",
     "cssnano": "^4.1.7",
@@ -79,7 +80,6 @@
     "js-yaml-loader": "^1.0.1",
     "marked": "^0.5.2",
     "mini-css-extract-plugin": "^0.4.5",
-    "ngx-auto-scroll": "^1.1.0",
     "ngx-bootstrap": "^3.1.2",
     "ngx-zone-scheduler": "^2.2.0",
     "node-sass": "^4.10.0",
@@ -91,7 +91,7 @@
     "postcss-preset-env": "^6.4.0",
     "preload-webpack-plugin": "3.0.0-beta.3",
     "raw-loader": "^0.5.1",
-    "sanitize-html": "^1.19.2",
+    "sanitize-html": "^1.19.3",
     "script-ext-html-webpack-plugin": "^2.1.3",
     "socket.io-client": "^2.2.0",
     "style-loader": "^0.23.1",
@@ -103,7 +103,7 @@
     "tslint-loader": "^3.6.0",
     "typescript": "<3.2.0",
     "url-loader": "^1.1.2",
-    "webpack": "4.26.1",
+    "webpack": "4.27.1",
     "webpack-bundle-analyzer": "^3.0.3",
     "webpack-clean-obsolete-chunks": "^0.4.0",
     "webpack-cli": "^3.1.2",
@@ -121,9 +121,9 @@
   },
   "dependencies": {
     "bcrypt-nodejs": "^0.0.3",
-    "bson": "^4.0.0",
+    "bson": "^4.0.1",
     "compression": "^1.7.3",
-    "core-js": "^2.5.7",
+    "core-js": "^2.6.0",
     "debug": "^4.1.0",
     "express": "^4.16.4",
     "express-csp": "^0.1.3",
@@ -157,12 +157,12 @@
     "@angular/upgrade": "^7.1.1",
     "@types/express": "^4.16.0",
     "@types/ioredis": "^4.0.4",
-    "@types/jquery": "^3.3.22",
-    "@types/lodash": "^4.14.118",
-    "@types/mongodb": "^3.1.15",
-    "@types/node": "^10.12.11",
+    "@types/jquery": "^3.3.23",
+    "@types/lodash": "^4.14.119",
+    "@types/mongodb": "^3.1.17",
+    "@types/node": "^10.12.12",
     "@types/serve-index": "^1.7.29",
-    "@types/socket.io": "^2.1.0",
+    "@types/socket.io": "^2.1.1",
     "@types/socket.io-client": "^1.4.32",
     "@types/socket.io-redis": "^1.0.25",
     "@types/uuid": "^3.4.4",
